代码提交

+++ src/main/java/service/AppmntsService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface AppmntsService {
 
     //新增
-    public Boolean addAppmnts(String userId, String objectId, int time, Date startDate,Date endDate);
+    public Boolean addAppmnts(String userId, String objectId, double time, Date startDate,Date endDate);
 
     //查询（按编号）
     public Appmnts getAppmntsById(String appmntsId);
@@ -20,9 +20,9 @@ public interface AppmntsService {
     public List<Appmnts> getAppmntsByUId(String userId);
 
     //更新预约信息
-    public Boolean updAppmnts(String appmntsId,int time, Date startDate,Date endDate,String appmntsAdder,String appmntsState);
+    public Boolean updAppmnts(String appmntsId,double time, Date startDate,Date endDate,String appmntsAdder,String appmntsState);
 
     //更新评价
-    public Boolean updAppraise(String appmntsId,int appraise,String appContext);
+    public Boolean updAppraise(String appmntsId,double appraise,String appContext);
 
 }
