+++ src/main/java/serviceimpl/AppmntsServiceImpl.java
@@ -16,7 +16,7 @@ public class AppmntsServiceImpl implements AppmntsService {
     AppmntsDao appmntsDao;
 
     @Override
-    public Boolean addAppmnts(String userId, String objectId, int time, Date startDate, Date endDate){
+    public Boolean addAppmnts(String userId, String objectId, double time, Date startDate, Date endDate){
         return appmntsDao.addAppmnts(userId,objectId,time,startDate,endDate);
     }
 
@@ -36,12 +36,12 @@ public class AppmntsServiceImpl implements AppmntsService {
     }
 
     @Override
-    public Boolean updAppmnts(String appmntsId,int time, Date startDate,Date endDate,String appmntsAdder,String appmntsState){
+    public Boolean updAppmnts(String appmntsId,double time, Date startDate,Date endDate,String appmntsAdder,String appmntsState){
         return appmntsDao.updAppmnts(appmntsId,time,startDate,endDate,appmntsAdder,appmntsState);
     }
 
     @Override
-    public Boolean updAppraise(String appmntsId,int appraise,String appContext){
+    public Boolean updAppraise(String appmntsId,double appraise,String appContext){
         return appmntsDao.updAppraise(appmntsId,appraise,appContext);
     }
